feat(ci): add GitHub Actions
Pushes will now be verified to be compliant with the styling and linters
as configured through pre-commit, when run inside a MR this will
automatically be scoped to that specific MR only.

Further more on tag events release binaries will now be build for every
available target and uploaded as a release for download.